Use new syntax from hanami-cli. Make CLI commands capable of resolving their own Hanami projects dependencies
