+++ lib/hanami/cli/commands/assets.rb
@@ -1,13 +1,11 @@
 module Hanami
   module Cli
     module Commands
-      # FIXME: this must be a module
-      class Assets
-        include Hanami::Cli::Command
-        register "assets", subcommand: true
-
+      module Assets
         require "hanami/cli/commands/assets/precompile"
       end
     end
+
+    register "assets precompile", Commands::Assets::Precompile
   end
 end
