+++ lib/hanami/cli/commands/db.rb
@@ -1,11 +1,7 @@
 module Hanami
   module Cli
     module Commands
-      # FIXME: this must be a module
-      class Db
-        include Hanami::Cli::Command
-        register "db", subcommand: true
-
+      module Db
         require "hanami/cli/commands/db/version"
         require "hanami/cli/commands/db/create"
         require "hanami/cli/commands/db/drop"
@@ -14,6 +10,16 @@ module Hanami
         require "hanami/cli/commands/db/apply"
         require "hanami/cli/commands/db/console"
       end
+    end
+
+    register "db" do |prefix|
+      prefix.register "version", Commands::Db::Version
+      prefix.register "create",  Commands::Db::Create
+      prefix.register "drop",    Commands::Db::Drop
+      prefix.register "migrate", Commands::Db::Migrate
+      prefix.register "prepare", Commands::Db::Prepare
+      prefix.register "apply",   Commands::Db::Apply
+      prefix.register "console", Commands::Db::Console
     end
   end
 end
