+++ lib/hanami/cli/commands/db/apply.rb
@@ -1,11 +1,9 @@
 module Hanami
   module Cli
     module Commands
-      # FIXME: this must be a module
-      class Db
-        class Apply
-          include Hanami::Cli::Command
-          register "db apply"
+      module Db
+        class Apply < Command
+          requires "model.sql"
 
           def call(**options)
             context = Context.new(options: options)
@@ -16,9 +14,8 @@ module Hanami
           private
 
           def apply_migrations(context)
-            # FIXME: this should be unified here
-            require "hanami/commands/db/apply"
-            Hanami::Commands::DB::Apply.new({}).start
+            require "hanami/model/migrator"
+            Hanami::Model::Migrator.apply
           end
         end
       end
