+++ lib/hanami/cli/commands/generate.rb
@@ -1,11 +1,7 @@
 module Hanami
   module Cli
     module Commands
-      # FIXME: this must be a module
-      class Generate
-        include Hanami::Cli::Command
-        register "generate", subcommand: true
-
+      module Generate
         require "hanami/cli/commands/generate/app"
         require "hanami/cli/commands/generate/action"
         require "hanami/cli/commands/generate/mailer"
@@ -13,6 +9,15 @@ module Hanami
         require "hanami/cli/commands/generate/model"
         require "hanami/cli/commands/generate/secret"
       end
+    end
+
+    register "generate", aliases: ["g"] do |prefix|
+      prefix.register "app",       Commands::Generate::App
+      prefix.register "action",    Commands::Generate::Action
+      prefix.register "mailer",    Commands::Generate::Mailer
+      prefix.register "migration", Commands::Generate::Migration
+      prefix.register "model",     Commands::Generate::Model
+      prefix.register "secret",    Commands::Generate::Secret
     end
   end
 end
