--- lib/hanami/commands/apps.rb
@@ -1,4 +1,0 @@
-module Hanami
-  module Components
-  end
-end
