--- lib/hanami/commands/db/create.rb
@@ -1,19 +1,0 @@
-require 'hanami/commands/command'
-
-module Hanami
-  module Commands
-    # @api private
-    class DB
-      # @api private
-      class Create < Command
-        requires 'model.configuration'
-
-        # @api private
-        def start
-          require 'hanami/model/migrator'
-          Hanami::Model::Migrator.create
-        end
-      end
-    end
-  end
-end
